[🚚Rename]: 회원가입 페이지 체크박스 폴더 옮기기 1. 폴더를 atoms에 CheckBox로 옮김 2. 파일 구분하기 위해 옮겼습니다.
